Moved source files to src directory
